+++ services/finalizer/standard/parameters.go
@@ -23,6 +23,7 @@ import (
 	"github.com/wealdtech/chaind/services/chaindb"
 	"github.com/wealdtech/chaind/services/chaintime"
 	"github.com/wealdtech/chaind/services/metrics"
+	"golang.org/x/sync/semaphore"
 )
 
 type parameters struct {
@@ -33,6 +34,7 @@ type parameters struct {
 	chainTime        chaintime.Service
 	blocks           blocks.Service
 	finalityHandlers []handlers.FinalityHandler
+	activitySem      *semaphore.Weighted
 }
 
 // Parameter is the interface for service parameters.
@@ -95,6 +97,13 @@ func WithFinalityHandlers(handlers []handlers.FinalityHandler) Parameter {
 	})
 }
 
+// WithActivitySem sets the activity semaphore for this module.
+func WithActivitySem(sem *semaphore.Weighted) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.activitySem = sem
+	})
+}
+
 // parseAndCheckParameters parses and checks parameters to ensure that mandatory parameters are present and correct.
 func parseAndCheckParameters(params ...Parameter) (*parameters, error) {
 	parameters := parameters{
@@ -117,6 +126,9 @@ func parseAndCheckParameters(params ...Parameter) (*parameters, error) {
 	}
 	if parameters.blocks == nil {
 		return nil, errors.New("no blocks specified")
+	}
+	if parameters.activitySem == nil {
+		return nil, errors.New("no activity semaphore specified")
 	}
 
 	return &parameters, nil