+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - add configurable database connection pool size
   - separate read-only and read-write transactions internally
+  - share activity semaphore between blocks and finalizer modules
 
 0.6.10
   - avoid crash with uninitialised metrics